Separate Type Checker from Semantic Checker. Don't report type errors when type declarations are invalid.
